profile-service/requirements.txt
Flask==2.2.3
Flask-Cors==3.0.10
Flask-SQLAlchemy==3.0.3
Flask-JWT-Extended==4.5.2
psycopg2-binary==2.9.6
python-dotenv==1.0.0
requests==2.31.0
SQLAlchemy-Utils==0.41.1
cryptography==41.0.1
gunicorn==21.2.0
Werkzeug==2.2.3
marshmallow==3.19.0